Fixed install_script

+++ install_script.sh
@@ -12,6 +12,13 @@ function install_ss {
     chmod u+x ./install_server.sh
 }
 
+function install_ck_server {
+    # It needs for key generation in system
+    wget https://github.com/cbeuw/Cloak/releases/download/v2.7.0/ck-server-linux-amd64-v2.7.0 -O ck-server
+    chmod +x ck-server
+    mv ck-server /bin/ck-server # sudo permissions!
+}
+
 function clone_repo {
     echo "Repo cloning..."	
     git clone https://github.com/DobbyVPN/dobbyvpn-server.git
@@ -43,7 +50,8 @@ function replace_cloak_holders {
     # $4 - adminUID
     # $5 - domain-name (for RedirAddr)
     # $6 - cloak private key
-   
+    
+    cp ./dobbyvpn-server/cloak-server-template.conf ./dobbyvpn-server/cloak-server.conf
     sed -i "s/<keys-port>/$1/g" ./dobbyvpn-server/cloak-server.conf
     sed -i "s/<cloak-server-port>/$2/g" ./dobbyvpn-server/cloak-server.conf
     sed -i "s/<user-UID>/$3/g" ./dobbyvpn-server/cloak-server.conf
@@ -56,40 +64,50 @@ function save_credentials {
    # Function saves sensitive data to file
    # $1 - filename
    # $2 - associative array
-   local file=$1
-   local array=$2
+    local file=$1
+    local array=$2
 
-   echo "Saving credentials"
-   if [ -e "$file" ]; then
-     echo "$file already exists."
-     read -e -p "Do you want to override it?(Y/n): " choice
-     case "$choice" in
-	 y|Y)
-    	     ;;
-	 n|N)
-    	     ;;
-     esac	 	     
-
+    echo "Saving credentials"
+    if [ -e "$file" ]; then
+        echo "$file already exists."
+        read -e -p "Do you want to override it?(Y/n): " choice
+        case "$choice" in
+	        y|Y)
+                rm $file
+                for key in "${!array[@]}"; do
+                    echo "$key => ${array[$key]}" >> "$file"
+                done
+                return
+       	        ;;
+	        n|N)
+                return
+    	        ;;
+        esac	 	     
 	     
    fi
 
 
    for key in "${!array[@]}"; do
-     echo "$key => ${array[$key]}" >> "$file"
+    echo "$key => ${array[$key]}" >> "$file"
    done
 }
 
 function readArgs {
-   read -e -p "Enter Cloak Port: " -i 8443 CLOAK_PORT
-   read -e -p "Enter Api Port(outline): " -i 11111 OUTLINE_API_PORT
-   read -e -p "Enter Keys Port(outline): " -i 22222 OUTLINE_KEYS_PORT
-   read -e -p "Enter Domain Name: " DOMAIN_NAME
+    read -e -p "Enter Cloak Port: " -i 8443 CLOAK_PORT
+    read -e -p "Enter Api Port(outline): " -i 11111 OUTLINE_API_PORT
+    read -e -p "Enter Keys Port(outline): " -i 22222 OUTLINE_KEYS_PORT
+    read -e -p "Enter Domain Name: " DOMAIN_NAME
 
-   if [ -z "$DOMAIN_NAME" ]; then
-     echo "Error: you didn't enter domain name!" >&2
-     exit 1
-   fi
+    if [ -z "$DOMAIN_NAME" ]; then
+        echo "Error: you didn't enter domain name!" >&2
+        exit 1
+    fi
 
+    KEYPAIRS=$(bin/ck-server -key)
+    CLOAK_PRIVATE_KEY=$(echo $KEYPAIRS | cut -d" " -f13)
+    CLOAK_PUBLIC_KEY=$(echo $KEYPAIRS | cut -d" " -f5)
+    USER_UID=$(bin/ck-server -uid | cut -d" " -f4)
+    ADMIN_UID=$(bin/ck-server -uid | cut -d" " -f4)
 }
 
 function main {
@@ -104,7 +122,7 @@ function main {
 
     URL=$(generate_url)
     replace_caddy_holders $DOMAIN_NAME $URL $CLOAK_PORT
-    replace_cloak_holders $OUTLINE_KEYS_PORT $CLOAK_PORT #BypassUID $ADMIN_UID $DOMAIN_NAME $CLOAK_PRIVATE_KEY
+    replace_cloak_holders $OUTLINE_KEYS_PORT $CLOAK_PORT $USER_UID $ADMIN_UID $DOMAIN_NAME $CLOAK_PRIVATE_KEY
 
     docker-compose -f ./dobbyvpn-server/docker-compose.yaml up -d
 
@@ -112,6 +130,8 @@ function main {
     array_creds["Special-url"]=$URL
     array_creds["Cloak-public-key"]=$CLOAK_PUBLIC_KEY
     array_creds["Cloak-private-key"]=$CLOAK_PRIVATE_KEY
+    array_creds["User-uid"]=$USER_UID
+    array_creds["Admin-uid"]=$ADMIN_UID
     save_credentials $filename $array_creds
 
     echo "All credentials are saved in $filename"
